backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-bcrypt==1.0.1
pymongo==4.6.0
PyJWT==2.8.0
python-dotenv==1.0.0
